+++ packages/backend/src/tokens/interstitial.ts
@@ -4,10 +4,14 @@ import { API_VERSION, USER_AGENT } from '../constants';
 // DO NOT CHANGE: Runtime needs to be imported as a default export so that we can stub its dependencies with Sinon.js
 // For more information refer to https://sinonjs.org/how-to/stub-dependency/
 import runtime from '../runtime';
-import { callWithRetry } from '../util/callWithRetry';
-import { isStaging } from '../util/instance';
-import { isDevOrStagingUrl } from '../util/isDevOrStagingUrl';
-import { parsePublishableKey } from '../util/parsePublishableKey';
+import {
+  addClerkPrefix,
+  callWithRetry,
+  getClerkJsMajorVersionOrTag,
+  getScriptUrl,
+  isDevOrStagingUrl,
+  parsePublishableKey,
+} from '../shared';
 import { joinPaths } from '../util/path';
 import { TokenVerificationError, TokenVerificationErrorAction, TokenVerificationErrorReason } from './errors';
 import type { DebugRequestSate } from './request';
@@ -27,24 +31,6 @@ export type LoadInterstitialOptions = {
   isSatellite?: boolean;
   signInUrl?: string;
 } & MultiDomainAndOrProxyPrimitives;
-
-// TODO: use the same function from @clerk/shared once treeshakable
-export function addClerkPrefix(str: string | undefined) {
-  if (!str) {
-    return '';
-  }
-  let regex;
-  if (str.match(/^(clerk\.)+\w*$/)) {
-    regex = /(clerk\.)*(?=clerk\.)/;
-  } else if (str.match(/\.clerk.accounts/)) {
-    return str;
-  } else {
-    regex = /^(clerk\.)*/gi;
-  }
-
-  const stripped = str.replace(regex, '');
-  return `clerk.${stripped}`;
-}
 
 export function loadInterstitialFromLocal(options: Omit<LoadInterstitialOptions, 'apiUrl'>) {
   options.frontendApi = parsePublishableKey(options.publishableKey)?.frontendApi || options.frontendApi || '';
@@ -194,29 +180,3 @@ export function buildPublicInterstitialUrl(options: LoadInterstitialOptions) {
 
   return url.href;
 }
-
-// TODO: Move to @clerk/shared as the same logic is used in @clerk/react
-const getClerkJsMajorVersionOrTag = (frontendApi: string, pkgVersion?: string) => {
-  if (!pkgVersion && isStaging(frontendApi)) {
-    return 'staging';
-  }
-
-  if (!pkgVersion) {
-    return 'latest';
-  }
-
-  if (pkgVersion.includes('next')) {
-    return 'next';
-  }
-
-  return pkgVersion.split('.')[0] || 'latest';
-};
-
-const getScriptUrl = (
-  frontendApi: string,
-  { pkgVersion, clerkJSVersion }: { pkgVersion?: string; clerkJSVersion?: string },
-) => {
-  const noSchemeFrontendApi = frontendApi.replace(/http(s)?:\/\//, '');
-  const major = getClerkJsMajorVersionOrTag(frontendApi, pkgVersion);
-  return `https://${noSchemeFrontendApi}/npm/@clerk/clerk-js@${clerkJSVersion || major}/dist/clerk.browser.js`;
-};
